+++ src/main/java/org/folio/search/service/consortium/ConsortiumInstanceSearchService.java
@@ -1,9 +1,14 @@
 package org.folio.search.service.consortium;
 
+import static org.apache.commons.collections4.CollectionUtils.containsAny;
 import static org.apache.commons.collections4.CollectionUtils.isEmpty;
+import static org.folio.search.converter.ConsortiumHoldingMapper.toConsortiumHolding;
+import static org.folio.search.converter.ConsortiumItemMapper.toConsortiumItem;
 import static org.folio.search.service.SearchService.DEFAULT_MAX_SEARCH_RESULT_WINDOW;
-import static org.folio.search.utils.SearchUtils.INSTANCE_HOLDING_FIELD_NAME;
-import static org.folio.search.utils.SearchUtils.INSTANCE_ITEM_FIELD_NAME;
+import static org.folio.search.utils.IdentifierUtils.getHoldingIdentifierValue;
+import static org.folio.search.utils.IdentifierUtils.getHoldingTargetField;
+import static org.folio.search.utils.IdentifierUtils.getItemIdentifierValue;
+import static org.folio.search.utils.IdentifierUtils.getItemTargetField;
 import static org.opensearch.index.query.QueryBuilders.boolQuery;
 import static org.opensearch.index.query.QueryBuilders.termsQuery;
 import static org.opensearch.search.sort.SortBuilders.fieldSort;
@@ -13,14 +18,10 @@ import java.util.ArrayList;
 import java.util.List;
 import java.util.Objects;
 import java.util.Set;
-import java.util.UUID;
-import java.util.function.BiFunction;
-import java.util.stream.Collectors;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.log4j.Log4j2;
 import org.folio.search.configuration.properties.SearchConfigurationProperties;
-import org.folio.search.converter.ConsortiumHoldingMapper;
-import org.folio.search.converter.ConsortiumItemMapper;
+import org.folio.search.domain.dto.BatchIdsDto.IdentifierTypeEnum;
 import org.folio.search.domain.dto.ConsortiumHolding;
 import org.folio.search.domain.dto.ConsortiumHoldingCollection;
 import org.folio.search.domain.dto.ConsortiumItem;
@@ -67,7 +68,7 @@ public class ConsortiumInstanceSearchService {
       return new ConsortiumHolding();
     }
 
-    return ConsortiumHoldingMapper.toConsortiumHolding(instance.getId(), holding);
+    return toConsortiumHolding(instance.getId(), holding);
   }
 
   public ConsortiumItem getConsortiumItem(String id, CqlSearchRequest<Instance> searchRequest) {
@@ -86,16 +87,16 @@ public class ConsortiumInstanceSearchService {
       return new ConsortiumItem();
     }
 
-    return ConsortiumItemMapper.toConsortiumItem(instance.getId(), item);
+    return toConsortiumItem(instance.getId(), item);
   }
 
-  public ConsortiumHoldingCollection fetchConsortiumBatchHoldings(String tenant, Set<UUID> holdingIds) {
-    validateIdsCount(holdingIds.size());
+  public ConsortiumHoldingCollection fetchConsortiumBatchHoldings(String tenant,
+                                                                  Set<String> identifierValues,
+                                                                  IdentifierTypeEnum identifierType) {
+    validateIdsCount(identifierValues.size());
 
-    var ids = holdingIds.stream().map(UUID::toString).collect(Collectors.toSet());
-    var targetField = INSTANCE_HOLDING_FIELD_NAME + ".id";
-
-    var searchRecords = getConsortiumBatchResults(tenant, ids, targetField, this::mapToConsortiumHolding);
+    var searchRecords = getConsortiumBatchResults(tenant,
+      identifierType, identifierValues, getHoldingTargetField(identifierType), this::mapToConsortiumHolding);
 
     if (searchRecords.isEmpty()) {
       return new ConsortiumHoldingCollection();
@@ -103,16 +104,16 @@ public class ConsortiumInstanceSearchService {
 
     return new ConsortiumHoldingCollection()
       .holdings(searchRecords.stream().map(SearchResult::getRecords).flatMap(List::stream).toList())
-      .totalRecords(searchRecords.iterator().next().getTotalRecords());
+      .totalRecords(searchRecords.stream().map(SearchResult::getTotalRecords).mapToInt(i -> i).sum());
   }
 
-  public ConsortiumItemCollection fetchConsortiumBatchItems(String tenant, Set<UUID> itemIds) {
-    validateIdsCount(itemIds.size());
+  public ConsortiumItemCollection fetchConsortiumBatchItems(String tenant,
+                                                            Set<String> identifierValues,
+                                                            IdentifierTypeEnum identifierType) {
+    validateIdsCount(identifierValues.size());
 
-    var ids = itemIds.stream().map(UUID::toString).collect(Collectors.toSet());
-    var targetField = INSTANCE_ITEM_FIELD_NAME + ".id";
-
-    var searchRecords = getConsortiumBatchResults(tenant, ids, targetField, this::mapToConsortiumItem);
+    var searchRecords = getConsortiumBatchResults(tenant,
+      identifierType, identifierValues, getItemTargetField(identifierType), this::mapToConsortiumItem);
 
     if (searchRecords.isEmpty()) {
       return new ConsortiumItemCollection();
@@ -120,19 +121,23 @@ public class ConsortiumInstanceSearchService {
 
     return new ConsortiumItemCollection()
       .items(searchRecords.stream().map(SearchResult::getRecords).flatMap(List::stream).toList())
-      .totalRecords(searchRecords.iterator().next().getTotalRecords());
+      .totalRecords(searchRecords.stream().map(SearchResult::getTotalRecords).mapToInt(i -> i).sum());
   }
 
-  private <T> List<SearchResult<T>> getConsortiumBatchResults(String tenant, Set<String> ids, String targetField,
-                                                              BiFunction<Instance, Set<String>, List<T>> recordMapper) {
+  private <T> List<SearchResult<T>> getConsortiumBatchResults(String tenant,
+                                                              IdentifierTypeEnum identifierType,
+                                                              Set<String> identifierValues,
+                                                              String targetField,
+                                                              Mapper<Instance, IdentifierTypeEnum, Set<String>, List<T>>
+                                                                recordMapper) {
     var request = CqlSearchRequest.of(Instance.class, tenant, "", 0, 0, true, false, true);
-    var termsQuery = termsQuery(targetField, ids);
+    var termsQuery = termsQuery(targetField, identifierValues);
 
-    if (ids.size() < DEFAULT_MAX_SEARCH_RESULT_WINDOW) {
-      var searchSourceBuilder = queryBuilder(termsQuery, ids.size());
+    if (identifierValues.size() < DEFAULT_MAX_SEARCH_RESULT_WINDOW) {
+      var searchSourceBuilder = queryBuilder(termsQuery, identifierValues.size());
       var response = searchRepository.search(request, searchSourceBuilder);
       var searchResult = documentConverter.convertToSearchResult(response, request.getResourceClass(),
-        (hits, item) -> recordMapper.apply(item, ids));
+        (hits, item) -> recordMapper.apply(item, identifierType, identifierValues));
       var records = searchResult.getRecords().stream()
         .flatMap(List::stream)
         .toList();
@@ -147,7 +152,7 @@ public class ConsortiumInstanceSearchService {
 
     while (response.getHits() != null && response.getHits().getHits().length > 0) {
       var searchResult = documentConverter.convertToSearchResult(response, request.getResourceClass(),
-        (hits, item) -> recordMapper.apply(item, ids));
+        (hits, item) -> recordMapper.apply(item, identifierType, identifierValues));
       var records = searchResult.getRecords().stream()
         .flatMap(List::stream)
         .toList();
@@ -161,17 +166,37 @@ public class ConsortiumInstanceSearchService {
     return searchRecords;
   }
 
-  private List<ConsortiumHolding> mapToConsortiumHolding(Instance instance, Set<String> ids) {
-    return instance.getHoldings().stream()
-      .filter(holding -> ids.contains(holding.getId()))
-      .map(holding -> ConsortiumHoldingMapper.toConsortiumHolding(instance.getId(), holding))
-      .toList();
+  private List<ConsortiumHolding> mapToConsortiumHolding(Instance instance,
+                                                         IdentifierTypeEnum identifierType,
+                                                         Set<String> identifierValues) {
+    if (identifierType == IdentifierTypeEnum.ITEMBARCODE) {
+      return instance.getItems().stream()
+        .filter(item -> identifierValues.contains(item.getBarcode()))
+        .flatMap(item -> instance.getHoldings().stream()
+          .filter(holding -> item.getHoldingsRecordId().equals(holding.getId()))
+        )
+        .map(holding -> toConsortiumHolding(instance.getId(), holding))
+        .distinct()
+        .toList();
+    } else if (identifierType == IdentifierTypeEnum.INSTANCEHRID) {
+      return instance.getHoldings()
+        .stream()
+        .map(holding -> toConsortiumHolding(instance.getId(), holding))
+        .toList();
+    } else {
+      return instance.getHoldings().stream()
+        .filter(holding -> containsAny(identifierValues, getHoldingIdentifierValue(identifierType, holding)))
+        .map(holding -> toConsortiumHolding(instance.getId(), holding))
+        .toList();
+    }
   }
 
-  private List<ConsortiumItem> mapToConsortiumItem(Instance instance, Set<String> ids) {
+  private List<ConsortiumItem> mapToConsortiumItem(Instance instance,
+                                                   IdentifierTypeEnum identifierType,
+                                                   Set<String> identifierValues) {
     return instance.getItems().stream()
-      .filter(item -> ids.contains(item.getId()))
-      .map(holding -> ConsortiumItemMapper.toConsortiumItem(instance.getId(), holding))
+      .filter(item -> containsAny(identifierValues, getItemIdentifierValue(identifierType, item)))
+      .map(holding -> toConsortiumItem(instance.getId(), holding))
       .toList();
   }
 
@@ -189,5 +214,10 @@ public class ConsortiumInstanceSearchService {
       throw new RequestValidationException("IDs array size exceeds the maximum allowed limit %s".formatted(idsLimit),
         "size", Long.toString(count));
     }
+  }
+
+  @FunctionalInterface
+  public interface Mapper<T, U, V, R> {
+    R apply(T t, U u, V v);
   }
 }