+++ src/test/java/org/folio/search/service/consortium/ConsortiumInstanceSearchServiceTest.java
@@ -1,6 +1,9 @@
 package org.folio.search.service.consortium;
 
+import static java.util.UUID.randomUUID;
 import static org.assertj.core.api.Assertions.assertThat;
+import static org.folio.search.converter.ConsortiumHoldingMapper.toConsortiumHolding;
+import static org.folio.search.converter.ConsortiumItemMapper.toConsortiumItem;
 import static org.folio.search.utils.TestConstants.CENTRAL_TENANT_ID;
 import static org.folio.search.utils.TestConstants.MEMBER_TENANT_ID;
 import static org.junit.jupiter.api.Assertions.assertThrows;
@@ -13,15 +16,13 @@ import static org.mockito.Mockito.when;
 
 import com.google.common.collect.Sets;
 import java.util.Collections;
+import java.util.HashSet;
 import java.util.List;
-import java.util.UUID;
-import java.util.function.BiFunction;
 import java.util.stream.Collectors;
 import java.util.stream.IntStream;
 import org.apache.lucene.search.TotalHits;
 import org.folio.search.configuration.properties.SearchConfigurationProperties;
-import org.folio.search.converter.ConsortiumHoldingMapper;
-import org.folio.search.converter.ConsortiumItemMapper;
+import org.folio.search.domain.dto.BatchIdsDto.IdentifierTypeEnum;
 import org.folio.search.domain.dto.ConsortiumHolding;
 import org.folio.search.domain.dto.ConsortiumHoldingCollection;
 import org.folio.search.domain.dto.ConsortiumItem;
@@ -59,10 +60,10 @@ class ConsortiumInstanceSearchServiceTest {
 
   @Test
   void getConsortiumHolding_positive() {
-    var instanceId = UUID.randomUUID().toString();
-    var id = UUID.randomUUID().toString();
+    var instanceId = randomUUID().toString();
+    var id = randomUUID().toString();
     var request = Mockito.<CqlSearchRequest<Instance>>mock();
-    var holdings = List.of(holding(UUID.randomUUID().toString(), CENTRAL_TENANT_ID), holding(id, MEMBER_TENANT_ID));
+    var holdings = List.of(holding(randomUUID().toString(), CENTRAL_TENANT_ID), holding(id, MEMBER_TENANT_ID));
     var searchResult = SearchResult.of(1, List.of(new Instance().id(instanceId).holdings(holdings)));
 
     when(searchService.search(request)).thenReturn(searchResult);
@@ -76,7 +77,7 @@ class ConsortiumInstanceSearchServiceTest {
 
   @Test
   void getConsortiumHolding_positive_noRecords() {
-    var id = UUID.randomUUID().toString();
+    var id = randomUUID().toString();
     var request = Mockito.<CqlSearchRequest<Instance>>mock();
     var searchResult = SearchResult.of(0, Collections.<Instance>emptyList());
 
@@ -89,8 +90,8 @@ class ConsortiumInstanceSearchServiceTest {
 
   @Test
   void getConsortiumHolding_positive_noHoldings() {
-    var instanceId = UUID.randomUUID().toString();
-    var id = UUID.randomUUID().toString();
+    var instanceId = randomUUID().toString();
+    var id = randomUUID().toString();
     var request = Mockito.<CqlSearchRequest<Instance>>mock();
     var holdings = Collections.<Holding>emptyList();
     var searchResult = SearchResult.of(1, List.of(new Instance().id(instanceId).holdings(holdings)));
@@ -104,10 +105,10 @@ class ConsortiumInstanceSearchServiceTest {
 
   @Test
   void getConsortiumHolding_positive_noDesiredHolding() {
-    var instanceId = UUID.randomUUID().toString();
-    var id = UUID.randomUUID().toString();
+    var instanceId = randomUUID().toString();
+    var id = randomUUID().toString();
     var request = Mockito.<CqlSearchRequest<Instance>>mock();
-    var holdings = List.of(holding(UUID.randomUUID().toString(), CENTRAL_TENANT_ID));
+    var holdings = List.of(holding(randomUUID().toString(), CENTRAL_TENANT_ID));
     var searchResult = SearchResult.of(1, List.of(new Instance().id(instanceId).holdings(holdings)));
 
     when(searchService.search(request)).thenReturn(searchResult);
@@ -119,10 +120,10 @@ class ConsortiumInstanceSearchServiceTest {
 
   @Test
   void getConsortiumItem_positive() {
-    var instanceId = UUID.randomUUID().toString();
-    var id = UUID.randomUUID().toString();
+    var instanceId = randomUUID().toString();
+    var id = randomUUID().toString();
     var request = Mockito.<CqlSearchRequest<Instance>>mock();
-    var items = List.of(item(UUID.randomUUID().toString(), CENTRAL_TENANT_ID), item(id, MEMBER_TENANT_ID));
+    var items = List.of(item(randomUUID().toString(), CENTRAL_TENANT_ID), item(id, MEMBER_TENANT_ID));
     var searchResult = SearchResult.of(1, List.of(new Instance().id(instanceId).items(items)));
 
     when(searchService.search(request)).thenReturn(searchResult);
@@ -137,7 +138,7 @@ class ConsortiumInstanceSearchServiceTest {
 
   @Test
   void getConsortiumItem_positive_noRecords() {
-    var id = UUID.randomUUID().toString();
+    var id = randomUUID().toString();
     var request = Mockito.<CqlSearchRequest<Instance>>mock();
     var searchResult = SearchResult.of(0, Collections.<Instance>emptyList());
 
@@ -150,8 +151,8 @@ class ConsortiumInstanceSearchServiceTest {
 
   @Test
   void getConsortiumItem_positive_noItems() {
-    var instanceId = UUID.randomUUID().toString();
-    var id = UUID.randomUUID().toString();
+    var instanceId = randomUUID().toString();
+    var id = randomUUID().toString();
     var request = Mockito.<CqlSearchRequest<Instance>>mock();
     var items = Collections.<Item>emptyList();
     var searchResult = SearchResult.of(1, List.of(new Instance().id(instanceId).items(items)));
@@ -165,10 +166,10 @@ class ConsortiumInstanceSearchServiceTest {
 
   @Test
   void getConsortiumItem_positive_noDesiredItem() {
-    var instanceId = UUID.randomUUID().toString();
-    var id = UUID.randomUUID().toString();
+    var instanceId = randomUUID().toString();
+    var id = randomUUID().toString();
     var request = Mockito.<CqlSearchRequest<Instance>>mock();
-    var items = List.of(item(UUID.randomUUID().toString(), CENTRAL_TENANT_ID));
+    var items = List.of(item(randomUUID().toString(), CENTRAL_TENANT_ID));
     var searchResult = SearchResult.of(1, List.of(new Instance().id(instanceId).items(items)));
 
     when(searchService.search(request)).thenReturn(searchResult);
@@ -181,11 +182,11 @@ class ConsortiumInstanceSearchServiceTest {
   @Test
   void fetchConsortiumBatchHoldings_negative_exceedsAllowedIdsCount() {
     when(properties.getMaxSearchBatchRequestIdsCount()).thenReturn(10L);
-    var ids = IntStream.range(1, 12).mapToObj(i -> UUID.randomUUID()).collect(Collectors.toSet());
+    var ids = IntStream.range(1, 12).mapToObj(i -> randomUUID().toString()).collect(Collectors.toSet());
 
     var ex =
       assertThrows(RequestValidationException.class,
-        () -> service.fetchConsortiumBatchHoldings(CENTRAL_TENANT_ID, ids));
+        () -> service.fetchConsortiumBatchHoldings(CENTRAL_TENANT_ID, ids, IdentifierTypeEnum.ID));
 
     assertThat(ex.getMessage()).isEqualTo("IDs array size exceeds the maximum allowed limit %s"
       .formatted(properties.getMaxSearchBatchRequestIdsCount()));
@@ -193,77 +194,67 @@ class ConsortiumInstanceSearchServiceTest {
 
   @Test
   void fetchConsortiumBatchHoldings_positive_notExceedsSearchResultWindow() {
-    var ids = List.of(UUID.randomUUID(), UUID.randomUUID());
-    var instancesHoldings = List.of(
-      List.of(holding(UUID.randomUUID().toString(), CENTRAL_TENANT_ID),
-        holding(ids.get(0).toString(), MEMBER_TENANT_ID)),
-      List.of(holding(ids.get(1).toString(), CENTRAL_TENANT_ID)),
-      List.of(holding(UUID.randomUUID().toString(), CENTRAL_TENANT_ID)));
-    var expectedConsortiumHoldings = instancesHoldings.subList(0, instancesHoldings.size() - 1).stream()
-      .map(holdings -> {
-        var holding = holdings.size() > 1 ? holdings.get(1) : holdings.get(0);
-        return ConsortiumHoldingMapper.toConsortiumHolding(UUID.randomUUID().toString(), holding);
-      }).toList();
+    var ids = List.of(randomUUID().toString(), randomUUID().toString());
+    var expectedConsortiumHoldings = List.of(
+      toConsortiumHolding(randomUUID().toString(), holding(ids.get(0), MEMBER_TENANT_ID)),
+      toConsortiumHolding(randomUUID().toString(), holding(ids.get(1), CENTRAL_TENANT_ID))
+    );
     when(properties.getMaxSearchBatchRequestIdsCount()).thenReturn(10L);
     when(searchRepository.search(any(CqlSearchRequest.class), any(SearchSourceBuilder.class)))
       .thenReturn(mock(SearchResponse.class));
-    when(documentConverter.convertToSearchResult(any(SearchResponse.class), eq(Instance.class), any(BiFunction.class)))
+    when(documentConverter.convertToSearchResult(any(SearchResponse.class), eq(Instance.class), any()))
       .thenReturn(SearchResult.of(2, List.of(expectedConsortiumHoldings)));
     var expected = new ConsortiumHoldingCollection()
       .holdings(expectedConsortiumHoldings)
       .totalRecords(2);
 
-    var result = service.fetchConsortiumBatchHoldings(CENTRAL_TENANT_ID, Sets.newHashSet(ids));
+    var result = service.fetchConsortiumBatchHoldings(CENTRAL_TENANT_ID, new HashSet<>(ids), IdentifierTypeEnum.ID);
 
     assertThat(result).isEqualTo(expected);
   }
 
   @Test
   void fetchConsortiumBatchHoldings_positive_exceedsSearchResultWindow() {
-    var ids = IntStream.range(1, 10002).mapToObj(i -> UUID.randomUUID()).toList();
-    var instancesHoldings = List.of(
-      List.of(holding(UUID.randomUUID().toString(), CENTRAL_TENANT_ID),
-        holding(ids.get(0).toString(), MEMBER_TENANT_ID)),
-      List.of(holding(ids.get(1).toString(), CENTRAL_TENANT_ID)),
-      List.of(holding(UUID.randomUUID().toString(), CENTRAL_TENANT_ID)));
-    var expectedConsortiumHoldings = instancesHoldings.subList(0, instancesHoldings.size() - 1).stream()
-      .map(holdings -> {
-        var holding = holdings.size() > 1 ? holdings.get(1) : holdings.get(0);
-        return ConsortiumHoldingMapper.toConsortiumHolding(UUID.randomUUID().toString(), holding);
-      }).toList();
+    var ids = IntStream.range(1, 10002).mapToObj(i -> randomUUID().toString()).toList();
+    var expectedHolding1 = toConsortiumHolding(randomUUID().toString(), holding(ids.get(0), MEMBER_TENANT_ID));
+    var expectedHolding2 = toConsortiumHolding(randomUUID().toString(), holding(ids.get(1), CENTRAL_TENANT_ID));
     var responseMock1 = mock(SearchResponse.class);
     var responseMock2 = mock(SearchResponse.class);
+    var responseMock3 = mock(SearchResponse.class);
     var hit = mock(SearchHit.class);
-    when(responseMock2.getHits()).thenReturn(SearchHits.empty());
     when(hit.getSortValues()).thenReturn(new Object[]{""});
+    when(responseMock2.getHits()).thenReturn(
+      new SearchHits(new SearchHit[]{hit}, new TotalHits(1, TotalHits.Relation.EQUAL_TO), 1.0f));
     when(responseMock1.getHits()).thenReturn(
       new SearchHits(new SearchHit[]{hit}, new TotalHits(1, TotalHits.Relation.EQUAL_TO), 1.0f));
+    when(responseMock3.getHits()).thenReturn(SearchHits.empty());
     when(properties.getMaxSearchBatchRequestIdsCount()).thenReturn(20000L);
-    when(properties.getSearchConsortiumRecordsPageSize()).thenReturn(2);
+    when(properties.getSearchConsortiumRecordsPageSize()).thenReturn(1);
     when(searchRepository.search(any(CqlSearchRequest.class), any(SearchSourceBuilder.class)))
-      .thenReturn(responseMock1).thenReturn(responseMock2);
-    when(documentConverter.convertToSearchResult(eq(responseMock1), eq(Instance.class), any(BiFunction.class)))
-      .thenReturn(SearchResult.of(2, List.of(expectedConsortiumHoldings)));
+      .thenReturn(responseMock1).thenReturn(responseMock2).thenReturn(responseMock3);
+    when(documentConverter.convertToSearchResult(eq(responseMock1), eq(Instance.class), any()))
+      .thenReturn(SearchResult.of(1, List.of(List.of(expectedHolding1))));
+    when(documentConverter.convertToSearchResult(eq(responseMock2), eq(Instance.class), any()))
+      .thenReturn(SearchResult.of(1, List.of(List.of(expectedHolding2))));
     var expected = new ConsortiumHoldingCollection()
-      .holdings(expectedConsortiumHoldings)
+      .holdings(List.of(expectedHolding1, expectedHolding2))
       .totalRecords(2);
 
-    var result = service.fetchConsortiumBatchHoldings(CENTRAL_TENANT_ID, Sets.newHashSet(ids));
+    var result = service.fetchConsortiumBatchHoldings(CENTRAL_TENANT_ID, Sets.newHashSet(ids), IdentifierTypeEnum.ID);
 
     assertThat(result).isEqualTo(expected);
-    verify(searchRepository, times(2))
-      .search(any(CqlSearchRequest.class), any(SearchSourceBuilder.class));
-    verify(documentConverter)
-      .convertToSearchResult(any(SearchResponse.class), eq(Instance.class), any(BiFunction.class));
+    verify(searchRepository, times(3)).search(any(CqlSearchRequest.class), any(SearchSourceBuilder.class));
+    verify(documentConverter, times(2))
+      .convertToSearchResult(any(SearchResponse.class), eq(Instance.class), any());
   }
 
   @Test
   void fetchConsortiumBatchItems_negative_exceedsAllowedIdsCount() {
     when(properties.getMaxSearchBatchRequestIdsCount()).thenReturn(10L);
-    var ids = IntStream.range(1, 12).mapToObj(i -> UUID.randomUUID()).collect(Collectors.toSet());
+    var ids = IntStream.range(1, 12).mapToObj(i -> randomUUID().toString()).collect(Collectors.toSet());
 
-    var ex =
-      assertThrows(RequestValidationException.class, () -> service.fetchConsortiumBatchItems(CENTRAL_TENANT_ID, ids));
+    var ex = assertThrows(RequestValidationException.class,
+        () -> service.fetchConsortiumBatchItems(CENTRAL_TENANT_ID, ids, IdentifierTypeEnum.ID));
 
     assertThat(ex.getMessage()).isEqualTo("IDs array size exceeds the maximum allowed limit %s"
       .formatted(properties.getMaxSearchBatchRequestIdsCount()));
@@ -271,44 +262,32 @@ class ConsortiumInstanceSearchServiceTest {
 
   @Test
   void fetchConsortiumBatchItems_positive_notExceedsSearchResultWindow() {
-    var ids = List.of(UUID.randomUUID(), UUID.randomUUID());
-    var instancesItems = List.of(
-      List.of(item(UUID.randomUUID().toString(), CENTRAL_TENANT_ID),
-        item(ids.get(0).toString(), MEMBER_TENANT_ID)),
-      List.of(item(ids.get(1).toString(), CENTRAL_TENANT_ID)),
-      List.of(item(UUID.randomUUID().toString(), CENTRAL_TENANT_ID)));
-    var expectedConsortiumItems = instancesItems.subList(0, instancesItems.size() - 1).stream()
-      .map(items -> {
-        var item = items.size() > 1 ? items.get(1) : items.get(0);
-        return ConsortiumItemMapper.toConsortiumItem(UUID.randomUUID().toString(), item);
-      }).toList();
+    var ids = List.of(randomUUID().toString(), randomUUID().toString());
+    var expectedConsortiumItems = List.of(
+      toConsortiumItem(randomUUID().toString(), item(ids.get(0), MEMBER_TENANT_ID)),
+      toConsortiumItem(randomUUID().toString(), item(ids.get(1), CENTRAL_TENANT_ID))
+    );
     when(properties.getMaxSearchBatchRequestIdsCount()).thenReturn(10L);
     when(searchRepository.search(any(CqlSearchRequest.class), any(SearchSourceBuilder.class)))
       .thenReturn(mock(SearchResponse.class));
-    when(documentConverter.convertToSearchResult(any(SearchResponse.class), eq(Instance.class), any(BiFunction.class)))
+    when(documentConverter.convertToSearchResult(any(SearchResponse.class), eq(Instance.class), any()))
       .thenReturn(SearchResult.of(2, List.of(expectedConsortiumItems)));
     var expected = new ConsortiumItemCollection()
       .items(expectedConsortiumItems)
       .totalRecords(2);
 
-    var result = service.fetchConsortiumBatchItems(CENTRAL_TENANT_ID, Sets.newHashSet(ids));
+    var result = service.fetchConsortiumBatchItems(CENTRAL_TENANT_ID, Sets.newHashSet(ids), IdentifierTypeEnum.ID);
 
     assertThat(result).isEqualTo(expected);
   }
 
   @Test
   void fetchConsortiumBatchItems_positive_exceedsSearchResultWindow() {
-    var ids = IntStream.range(1, 10002).mapToObj(i -> UUID.randomUUID()).toList();
-    var instancesItems = List.of(
-      List.of(item(UUID.randomUUID().toString(), CENTRAL_TENANT_ID),
-        item(ids.get(0).toString(), MEMBER_TENANT_ID)),
-      List.of(item(ids.get(1).toString(), CENTRAL_TENANT_ID)),
-      List.of(item(UUID.randomUUID().toString(), CENTRAL_TENANT_ID)));
-    var expectedConsortiumItems = instancesItems.subList(0, instancesItems.size() - 1).stream()
-      .map(items -> {
-        var item = items.size() > 1 ? items.get(1) : items.get(0);
-        return ConsortiumItemMapper.toConsortiumItem(UUID.randomUUID().toString(), item);
-      }).toList();
+    var ids = IntStream.range(1, 10002).mapToObj(i -> randomUUID().toString()).toList();
+    var expectedConsortiumItems = List.of(
+      toConsortiumItem(randomUUID().toString(), item(ids.get(0), MEMBER_TENANT_ID)),
+      toConsortiumItem(randomUUID().toString(), item(ids.get(1), CENTRAL_TENANT_ID))
+    );
     var responseMock1 = mock(SearchResponse.class);
     var responseMock2 = mock(SearchResponse.class);
     var hit = mock(SearchHit.class);
@@ -320,19 +299,19 @@ class ConsortiumInstanceSearchServiceTest {
     when(properties.getSearchConsortiumRecordsPageSize()).thenReturn(2);
     when(searchRepository.search(any(CqlSearchRequest.class), any(SearchSourceBuilder.class)))
       .thenReturn(responseMock1).thenReturn(responseMock2);
-    when(documentConverter.convertToSearchResult(eq(responseMock1), eq(Instance.class), any(BiFunction.class)))
+    when(documentConverter.convertToSearchResult(eq(responseMock1), eq(Instance.class), any()))
       .thenReturn(SearchResult.of(2, List.of(expectedConsortiumItems)));
     var expected = new ConsortiumItemCollection()
       .items(expectedConsortiumItems)
       .totalRecords(2);
 
-    var result = service.fetchConsortiumBatchItems(CENTRAL_TENANT_ID, Sets.newHashSet(ids));
+    var result = service.fetchConsortiumBatchItems(CENTRAL_TENANT_ID, Sets.newHashSet(ids), IdentifierTypeEnum.ID);
 
     assertThat(result).isEqualTo(expected);
     verify(searchRepository, times(2))
       .search(any(CqlSearchRequest.class), any(SearchSourceBuilder.class));
     verify(documentConverter)
-      .convertToSearchResult(any(SearchResponse.class), eq(Instance.class), any(BiFunction.class));
+      .convertToSearchResult(any(SearchResponse.class), eq(Instance.class), any());
   }
 
   private Holding holding(String id, String tenantId) {
@@ -345,6 +324,6 @@ class ConsortiumInstanceSearchServiceTest {
     return new Item()
       .id(id)
       .tenantId(tenantId)
-      .holdingsRecordId(UUID.randomUUID().toString());
+      .holdingsRecordId(randomUUID().toString());
   }
 }